==> 1.Explore_NSSdata.Rmd <==
---
title: "Summarise_NSS_Dataset"
author: "Danielle L. Buss"
date: "2024-03-01"
output: html_document
---

## Loading NSS database
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#Load data into global environment
library(ggplot2)
library(reshape2)
library(dplyr)
library(tidyr)
library(forcats)
load("~/Documents/4-Oceans/3.NSS/March2024/NSS_July2023.RData")
```

```{r, Load libraries, dataset and check headers, include=FALSE}
head(df)
theme_publish <- function(base_size = 12, base_family = "",
                          line_size = 0.25, ...) {
  half_line <- base_size / 2
  small_rel <- 0.8
  small_size <- small_rel * base_size
  
  theme_bw(base_size = base_size, base_family = base_family, ...) %+replace%
    theme(
      rect = element_rect(fill = "transparent", colour = NA, color = NA,
                          linewidth = 0, linetype = 0),
      text = element_text(family = base_family, face = "plain",
                          colour = "black", size = base_size, hjust = 0.5,
                          vjust = 0.5, angle = 0, lineheight = 0.9,
                          margin = ggplot2::margin(), debug = F),
      
      axis.text = element_text(size = small_size),
      axis.text.x = element_text(margin = ggplot2::margin(t = small_size/4),
                                 vjust = 1),
      axis.text.y = element_text(margin = ggplot2::margin(r = small_size/4), 
                                 hjust = 1),
      axis.title.x = element_text(margin = ggplot2::margin(t = small_size,
                                                           b = small_size)),
      axis.title.y = element_text(angle = 90,
                                  margin = ggplot2::margin(r = small_size,
                                                           l = small_size/4)),
      axis.ticks = element_line(colour = "black", linewidth = line_size),
      axis.ticks.length = unit(0.25, 'lines'),
      
      axis.line = element_line(colour = "black", linewidth = line_size),
      axis.line.x = element_line(colour = "black", linewidth = line_size), 
      axis.line.y = element_line(colour = "black", linewidth = line_size), 
      
      legend.spacing = unit(base_size/4, "pt"),
      legend.key = element_blank(),
      legend.key.size = unit(1 * base_size, "pt"),
      legend.key.width = unit(1.5 * base_size, 'pt'),
      legend.text = element_text(size = rel(small_rel)),
      legend.title = element_text(size = rel(small_rel), face = 'bold'),
      legend.position = 'bottom',
      legend.box = 'horizontal',
      
      panel.spacing = unit(1, "lines"),
      panel.background = element_blank(),
      panel.border = element_blank(), 
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      
      strip.text = element_text(size = base_size),
      strip.background = element_rect(fill = NA, colour = "black", linewidth = 0.125),
      strip.text.x = element_text(face = 'bold', hjust = 0,
                                  margin = ggplot2::margin(b = small_size/2,
                                                           t = small_size/4)),
      strip.text.y = element_text(angle = -90, face = 'bold',
                                  margin = ggplot2::margin(l = small_size/2,
                                                           r = small_size/4)),
      
      plot.margin = unit(c(5,5,0,0), "pt"),
      plot.background = element_blank(),
      plot.title = element_text(face = "bold", size = 1.2 * base_size, 
                                margin = ggplot2::margin(b = half_line),
                                hjust = 0)
    )
}
```

## Summarising NSS database:

### Number of countries
```{r, countries}
length(unique(df$Country))
unique(df$Country)
```

### Number of countries
```{r, sites}
length(unique(df$Site.name))
aggregate(Site.name ~ Country, data=unique(df[c("Site.name","Country")]), FUN=length)
```
### Number of assemblages
```{r, assemblages}
length(unique(df$Assemblage.or.sub.assemblage))
aggregate(Assemblage.or.sub.assemblage ~ Country, data=unique(df[c("Assemblage.or.sub.assemblage","Country")]), FUN=length)
```

#### When were NISP data reported by country 
```{r, Year.reported}
ggplot(df, aes(Year.reported, color=Country, fill=Country)) + geom_histogram(bins=8) + facet_wrap(~Country, nrow=2) + theme_publish() + 
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8) + theme(legend.position = "none")
```

#### Counts of urban and rural 
```{r, urban and rural }
aggregate(Assemblage.or.sub.assemblage ~ Rural.urban.or.neither, data=unique(df[c("Assemblage.or.sub.assemblage","Rural.urban.or.neither")]), FUN=length)
df_tmp<-aggregate(Assemblage.or.sub.assemblage ~ Rural.urban.or.neither + Country, data=unique(df[c("Assemblage.or.sub.assemblage","Rural.urban.or.neither", "Country")]), FUN=length)
ggplot(df_tmp, aes(Rural.urban.or.neither, Assemblage.or.sub.assemblage, fill=Country)) + geom_bar(stat="identity") + facet_wrap(~Country, nrow=2) + theme_publish() + 
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none")
```

#### Start.date to End.Date
```{r, Start and End Date, warnings = FALSE}
df_tmp<-unique(df[c("Assemblage.or.sub.assemblage","Start.date.CE", "End.date.CE","Country")])
df_tmp$Assemblage.or.sub.assemblage<-as.factor(df_tmp$Assemblage.or.sub.assemblage)
ggplot(df_tmp) +
  geom_segment(size=6,alpha=0.8, aes(x=Start.date.CE, xend=End.date.CE, color=Country, y=fct_reorder(Assemblage.or.sub.assemblage,
                                                                                                     Start.date.CE, min),
                                     yend=fct_reorder(Assemblage.or.sub.assemblage, Start.date.CE, min))) + facet_wrap(~Country) + labs(x="Year", y ="No. Assemblages") + theme_publish() + 
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        axis.ticks.y = element_line(size = 5),
        axis.title.x = element_text(size=18),
        plot.title = element_text(size=20),
        axis.text.x = element_text(size=16),
        strip.text = element_text(size=20)) +
  scale_color_brewer(type = "qual", palette = 8)  + theme(legend.position = "none")
```

#### Start.date to End.Date - < 400 yrs chronology
```{r, Start and End Date < 400 yrs}
df$time.range<-df$End.date.CE - df$Start.date.CE
df %>%
  filter(!is.na(time.range)) %>%
  filter(Start.date.CE > 0) %>%
  filter(End.date.CE > 0) %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  ggplot() +
  geom_segment(size=6,alpha=0.8, aes(x=Start.date.CE, xend=End.date.CE, color=Country, y=fct_reorder(Assemblage.or.sub.assemblage,
                                                                                                     Start.date.CE, min),
                                     yend=fct_reorder(Assemblage.or.sub.assemblage, Start.date.CE, min))) + facet_wrap(~Country) + labs(x="Year", y ="No. Assemblages") + theme_publish() + 
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        axis.ticks.y = element_line(size = 5),
        axis.title.x = element_text(size=18),
        plot.title = element_text(size=20),
        axis.text.x = element_text(size=16),
        strip.text = element_text(size=20)) +
  scale_color_brewer(type = "qual", palette = 8)  + theme(legend.position = "none")
```

#### Start.date to End.Date - < 300 yrs chronology
```{r, Start and End Date < 300 yrs}
df %>%
  filter(!is.na(time.range)) %>%
  filter(Start.date.CE > 0) %>%
  filter(End.date.CE > 0) %>%
  filter((End.date.CE - Start.date.CE) < 300) %>%
  ggplot() +
  geom_segment(size=6,alpha=0.8, aes(x=Start.date.CE, xend=End.date.CE, color=Country, y=fct_reorder(Assemblage.or.sub.assemblage,
                                                                                                     Start.date.CE, min),
                                     yend=fct_reorder(Assemblage.or.sub.assemblage, Start.date.CE, min))) + facet_wrap(~Country) + labs(x="Year", y ="No. Assemblages") + theme_publish() + 
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        axis.ticks.y = element_line(size = 5),
        axis.title.x = element_text(size=18),
        plot.title = element_text(size=20),
        axis.text.x = element_text(size=16),
        strip.text = element_text(size=20)) +
  scale_color_brewer(type = "qual", palette = 8)  + theme(legend.position = "none")
```

#### Start.date to End.Date - < 200 yrs chronology
```{r, Start and End Date < 200 yrs}
df %>%
  filter(!is.na(time.range)) %>%
  filter(Start.date.CE > 0) %>%
  filter(End.date.CE > 0) %>%
  filter((End.date.CE - Start.date.CE) < 200) %>%
  ggplot() +
  geom_segment(size=6,alpha=0.8, aes(x=Start.date.CE, xend=End.date.CE, color=Country, y=fct_reorder(Assemblage.or.sub.assemblage,
                                                                                                     Start.date.CE, min),
                                     yend=fct_reorder(Assemblage.or.sub.assemblage, Start.date.CE, min))) + facet_wrap(~Country) + labs(x="Year", y ="No. Assemblages") + theme_publish() + 
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        axis.ticks.y = element_line(size = 5),
        axis.title.x = element_text(size=18),
        plot.title = element_text(size=20),
        axis.text.x = element_text(size=16),
        strip.text = element_text(size=20)) +
  scale_color_brewer(type = "qual", palette = 8)  + theme(legend.position = "none")
```

```{r, Recovery method per country}
df_tmp<-aggregate(Assemblage.or.sub.assemblage ~ Recovery.method..hand.collected..sieved..both..unknown. + Country, data=unique(df[c("Assemblage.or.sub.assemblage","Recovery.method..hand.collected..sieved..both..unknown.", "Country")]), FUN=length)
ggplot(df_tmp, aes(Recovery.method..hand.collected..sieved..both..unknown., Assemblage.or.sub.assemblage, fill=Country)) + geom_bar(stat="identity") + facet_wrap(~Country, nrow=2) + theme_publish() + 
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                         axis.text.x = element_text(angle=45, hjust=1))
```

```{r, No. of unidentified specimens per country}
df_tmp<-aggregate(Number.of.unidentified.fish.specimens ~ Country, data=df, FUN=sum)
ggplot(df_tmp, aes(Number.of.unidentified.fish.specimens, Country, fill=Country)) + geom_bar(stat="identity") + theme_publish() +
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                         axis.text.x = element_text(angle=45, hjust=1))
```

```{r, No. of identified specimens per country}
df_tmp<-aggregate(Total.Fish.NISP ~ Country, data=df, FUN=sum)
ggplot(df_tmp, aes(Total.Fish.NISP, Country, fill=Country)) + geom_bar(stat="identity") + theme_publish() +
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                         axis.text.x = element_text(angle=45, hjust=1))
```
```{r, No. of identified specimens per red/amber/green per country}
df_tmp<-aggregate(Total.Fish.NISP ~ Data.quality..green.amber.or.red. + Country, data=df, FUN=sum)
ggplot(df_tmp, aes(Total.Fish.NISP, Country, fill=Country)) + geom_bar(stat="identity") + facet_wrap(~Data.quality..green.amber.or.red.) + theme_publish() +
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                         axis.text.x = element_text(angle=45, hjust=1))
```

#### Need to lump sieved & hand_collected data per assemblage here (NDY)

```{r, No. of identified specimens per red/amber/green per country <400 yrs}
df_tmp<-aggregate(Total.Fish.NISP ~ Data.quality..green.amber.or.red. + Country, data=df[df$time.range < 400,], FUN=sum)
ggplot(df_tmp, aes(Total.Fish.NISP, Country, fill=Country)) + geom_bar(stat="identity") + facet_wrap(~Data.quality..green.amber.or.red.) + theme_publish() +
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                         axis.text.x = element_text(angle=45, hjust=1))
```

```{r, No. of identified specimens per assemblage per country <400 yrs}
df_tmp<-aggregate(Total.Fish.NISP ~ Assemblage.or.sub.assemblage + Country, data=df[df$time.range < 400,], FUN=sum)
ggplot(df_tmp, aes(log(Total.Fish.NISP), color=Country, fill=Country)) + geom_histogram(bins=6) + facet_wrap(~Country) + theme_publish() +
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                         axis.text.x = element_text(angle=45, hjust=1))
```

```{r, Remove smallest and largest assemblages}
quantile(df$Total.Fish.NISP, 0.975)
quantile(df$Total.Fish.NISP, 0.025)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400)
ggplot(df_tmp, aes(Total.Fish.NISP, color=Country, fill=Country)) + geom_histogram(bins=6) + facet_wrap(~Country) + theme_publish() +
  scale_color_brewer(type = "qual", palette = 8) + 
  scale_fill_brewer(type = "qual", palette = 8)  + theme(legend.position = "none",
                                                      axis.text.x = element_text(angle=45, hjust=1))
```
###Counts by taxa (top 50)
```{r, Counts per taxa}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ GBIF_species, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 5.9)
ggplot(df_tmp, aes(log(NISP_count))) + geom_histogram(bins=50)
df_tmp<-df_tmp %>% filter(!GBIF_species == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(GBIF_species, NISP_count), fill=GBIF_species)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Species")
```
###Counts by taxa (excluding taxa with low counts)
```{r, Counts per genus}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ GBIF_genus, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 5.9)
df_tmp<-df_tmp %>% filter(!GBIF_genus == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(GBIF_genus, NISP_count), fill=GBIF_genus)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Genus")
```
```{r, Counts per family top 30}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ GBIF_family, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 5.9)
df_tmp<-df_tmp %>% filter(!GBIF_family == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(GBIF_family, NISP_count), fill=GBIF_family)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Genus")
```
```{r, Counts per order top 20}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ GBIF_order, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 4)
df_tmp<-df_tmp %>% filter(!GBIF_order == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(GBIF_order, NISP_count), fill=GBIF_order)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Genus")
```
###Counts by traits
```{r, Counts per traits - habitat}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ Trait_Habitat, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 4)
df_tmp<-df_tmp %>% filter(!Trait_Habitat == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(Trait_Habitat, NISP_count), fill=Trait_Habitat)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Habitat")
```

```{r, Counts per traits - foodweb}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ Trait_Foodweb, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 4)
df_tmp<-df_tmp %>% filter(!Trait_Foodweb == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(Trait_Foodweb, NISP_count), fill=Trait_Foodweb)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Foodweb")
```
```{r, Counts per traits - life history strategy}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ Trait_LifeHistory, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 4)
df_tmp<-df_tmp %>% filter(!Trait_LifeHistory == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(Trait_LifeHistory, NISP_count), fill=Trait_LifeHistory)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Life history strategy")
```
```{r, Counts per traits - temp range}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp$temp.range<-as.numeric(df_tmp$Trait_Temp_Max) - as.numeric(df_tmp$Trait_Temp_Min)
df_tmp<-aggregate(NISP_count ~ temp.range, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 7)
df_tmp<-df_tmp %>% filter(!temp.range == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(temp.range, as.numeric(temp.range)), fill=temp.range)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Temp range")
```
```{r, Counts per traits - temp midpoint}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp$temp.midpoint<-as.numeric(df_tmp$Trait_Temp_Max) - ((as.numeric(df_tmp$Trait_Temp_Max) - as.numeric(df_tmp$Trait_Temp_Min))/2)
df_tmp<-aggregate(NISP_count ~ temp.midpoint, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 7)
df_tmp<-df_tmp %>% filter(!temp.midpoint == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(temp.midpoint, as.numeric(temp.midpoint)), fill=temp.midpoint)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Temp midpoint")
```
```{r, Counts per traits - temp min}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp$Trait_Temp_Min<-as.numeric(df_tmp$Trait_Temp_Min)
df_tmp<-aggregate(NISP_count ~ Trait_Temp_Min, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 7)
df_tmp<-df_tmp %>% filter(!Trait_Temp_Min == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(Trait_Temp_Min, Trait_Temp_Min), fill=Trait_Temp_Min)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Temp range min")
```
```{r, Counts per traits - temp max}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp$Trait_Temp_Max<-as.numeric(df_tmp$Trait_Temp_Max)
df_tmp<-aggregate(NISP_count ~ Trait_Temp_Max, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 7)
df_tmp<-df_tmp %>% filter(!Trait_Temp_Max == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(Trait_Temp_Max, Trait_Temp_Max), fill=Trait_Temp_Max)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Temp range max")
```

```{r, Counts per traits - endangerment status}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ Trait_endangered, data=df_tmp, FUN=sum)
df_tmp<-df_tmp %>% filter(log(NISP_count) > 4)
df_tmp<-df_tmp %>% filter(!Trait_endangered == "NA")
ggplot(df_tmp, aes(NISP_count, reorder(Trait_endangered, NISP_count), fill=Trait_endangered)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="Total number of identified specimens", y="Genus")
```
```{r, Critically endangered species/counts/country}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
#df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude)) %>%
  filter(Trait_endangered == "Critically endangered")
df_tmp$all<-"All"
df_tmp<-aggregate(NISP_count ~ Country + GBIF_species, data=df_tmp, FUN=sum)
#df_tmp<-df_tmp %>% filter(log(NISP_count) > 4)
df_tmp<-df_tmp %>% filter(!GBIF_species == "NA")
ggplot(df_tmp, aes(log(NISP_count), reorder(Country, NISP_count), fill=GBIF_species)) + geom_bar(stat="identity") + theme_publish() + theme(legend.position = "none") + labs(x="log(Total number of identified specimens)", y="Species") + facet_wrap(~GBIF_species)
```

###By Ocean
```{r, Summary data by "Sea"}
rm(df_tmp)
df_tmp<-df %>% filter(Total.Fish.NISP < quantile(df$Total.Fish.NISP, 0.975))
df_tmp<-df_tmp %>% filter(Total.Fish.NISP > quantile(df$Total.Fish.NISP, 0.025)) 
df_tmp<- df_tmp %>%
  filter((End.date.CE - Start.date.CE) < 400) %>%
  filter(complete.cases(Decimal.Longitude, Decimal.Latitude))
df_tmp$all<-"All"
a<-df_tmp %>% mutate(binlon = cut(Decimal.Longitude, seq(from = min(Decimal.Longitude), to = max(Decimal.Longitude), by = 10), include.lowest = T, right = F),
           binlat = cut(Decimal.Latitude, seq(from = min(Decimal.Latitude), to = max(Decimal.Latitude), by = 10), include.lowest = T, right = F)) %>% 
    group_by(all, binlat, binlon) %>% 
    summarise(count = sum(Total.Fish.NISP))
df$Ocean<-recode(df$Country)

```
